feat(api): added an alternative filter (default filter is for javascript)
