+++ server/api/index.js
@@ -4,6 +4,7 @@ module.exports = router
 router.use('/users', require('./users'))
 router.use('/country', require('./country'))
 router.use('/state', require('./state'))
+router.use('/city', require('./city'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
